Feat: api 서비스 주문

+++ src/main/kotlin/com/hwalaon/wezxro_server/domain/wapi/controller/response/AddOrderResponse.kt
@@ -1,3 +1,3 @@
 package com.hwalaon.wezxro_server.domain.wapi.controller.response
 
-data class AddOrderResponse(val order: Int)
+data class AddOrderResponse(val order: Long)
